fix: fix error handling in Makefile
Makefile "ignores" errors inside for loop.  It only looks at the exit
code of the for loop itself which is the exit code of the last run
command inside the for loop.  Or in other words: Makefile "ignores"
errors inside for loop except for the last one.

Add `set -e` so that Makefile correctly fails whenever any command fails
inside for loop and returns error code.